requirements.txt
boto3
jinja2
python-box>=5.4.1
attrs>=21.4.0
attrs_mate>=1.0.2
loggerFactory==0.0.5
colorama
